feat: move the function getActualNumValue to util/space
