+++ packages/vchart/src/component/indicator/indicator.ts
@@ -8,7 +8,7 @@ import { BaseComponent } from '../base';
 import type { IRegion } from '../../region/interface';
 import type { IIndicator, IIndicatorItemSpec, IIndicatorTheme } from './interface';
 import type { Maybe } from '../../typings';
-import { isValid, isFunction, array, merge, eachSeries } from '../../util';
+import { isValid, isFunction, array, merge, eachSeries, transformToGraphic, getActualNumValue } from '../../util';
 import { isEqual } from '@visactor/vutils';
 import { indicatorMapper } from './util';
 import type { IModel } from '../../model/interface';
@@ -18,9 +18,7 @@ import { Indicator as IndicatorComponents } from '@visactor/vrender-components';
 // eslint-disable-next-line no-duplicate-imports
 import type { IndicatorAttributes } from '@visactor/vrender-components';
 import type { IGraphic, INode } from '@visactor/vrender';
-import { transformToGraphic } from '../../util/style';
 import type { IVisualScale, IVisualSpecStyle, VisualType, FunctionType } from '../../typings/visual';
-import { getActualNumValue } from '../../series/util/utils';
 
 export class Indicator extends BaseComponent implements IIndicator {
   static speckey = 'indicator';
