+++ packages/vchart/src/series/bar/bar.ts
@@ -4,9 +4,8 @@ import { Direction } from '../../typings/space';
 import { CartesianSeries } from '../cartesian/cartesian';
 import { MarkTypeEnum } from '../../mark/interface';
 import { AttributeLevel } from '../../constant';
-import { getActualNumValue } from '../util/utils';
 import type { Maybe, Datum, DirectionType } from '../../typings';
-import { merge, valueInScaleRange } from '../../util';
+import { merge, valueInScaleRange, getActualNumValue } from '../../util';
 import type { BarAppearPreset, IBarAnimationParams } from './animation';
 import { animationConfig, shouldDoMorph, userAnimationConfig } from '../../animation/utils';
 import type { IBarSeriesSpec, IBarSeriesTheme } from './interface';
